Extrapolated API methods used by tft.js into util file
