+++ nacho-bot/commands/tft.js
@@ -5,11 +5,7 @@ Teamfight Tactics: summoner history, upcoming matches, using TeemoJS NPM Package
 //name, icon, level, rank, winrate
 
 const { MessageEmbed } = require('discord.js');
-const TeemoJS = require('teemojs');
-const request = require('request');
-const { prefix, token, mikesRiotAPIKey } = require('../config.json');
-const { match } = require('assert');
-let teemo = TeemoJS(mikesRiotAPIKey);
+const tftApi = require ('../scripts/tftApi.js');
 
 async function main(message, args) {
     if (args.length == 0) {
@@ -32,7 +28,7 @@ async function main(message, args) {
         }
     }
     
-    const summoner = await getSummonerData(summonerName);
+    const summoner = await tftApi.getSummonerData(summonerName);
 
     if (summoner === undefined) {
         const errorEmbed = new MessageEmbed()
@@ -59,42 +55,6 @@ async function main(message, args) {
     summonerDataEmbed.addField("Rank", summoner.rank);
     summonerDataEmbed.addField("Winrate", summoner.winrate);
     message.reply(summonerDataEmbed);
-}
-
-async function getSummonerData(summonerName) {
-    let output = {};
-
-    const summoner = await getSummonerByName(summonerName);
-    if (summoner === null) {
-        return undefined;
-    }
-    const entries = await getLeagueEntries(summoner.id);//UtcMpsT00sv70OmKiRsdZY9mJTBOF7tjhvczP9zIzoHw80w
-    const entry = entries[0];
-
-    output.name = summoner.name;
-    output.icon = summoner.profileIconId;
-    output.level = summoner.summonerLevel;
-
-    if (entry === undefined) {
-        output.rank = "unranked scrub";
-        output.lp = "-1";
-        output.winrate = "probably terrible";
-    }
-    else {
-        output.rank = `${entry.tier} - ${entry.rank}`;
-        output.lp = entry.leaguePoints;
-        const winrateCalc = (parseInt(entry.wins) / parseInt(entry.losses) * 100).toFixed(2);
-        output.winrate = `${entry.wins}/${entry.losses} (${winrateCalc}%)`;
-    }
-    return output;
-}
-
-async function getSummonerByName(summonerName) {
-    return await teemo.get('na1', 'tftSummoner.getBySummonerName', summonerName);
-}
-
-async function getLeagueEntries(summonerId) {
-    return await teemo.get('na1', 'tftLeague.getLeagueEntriesForSummoner', summonerId);
 }
 
 module.exports = {
